Refactor loading actions
