--- client/modules/IDE/actions/loader.js
@@ -1,9 +1,0 @@
-import * as ActionTypes from '../../../constants';
-
-export function startLoader() {
-  return { type: ActionTypes.START_LOADING };
-}
-
-export function stopLoader() {
-  return { type: ActionTypes.STOP_LOADING };
-}
